requirements.txt
openai-whisper
pytube
gradio